# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ReHub API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List Users",
        "description": "Admin only: List all users.",
        "operationId": "list_users_api_v1_admin_users_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserMe"
                  },
                  "title": "Response List Users Api V1 Admin Users Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/status": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Update User Account Status",
        "description": "Admin only: Ban or unban a user by toggling is_active.",
        "operationId": "update_user_account_status_api_v1_admin_users__user_id__status_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMe"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/listings/pending": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get Pending Listings Route",
        "description": "Admin only: List all pending listings.",
        "operationId": "get_pending_listings_route_api_v1_admin_listings_pending_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListingRead"
                  },
                  "title": "Response Get Pending Listings Route Api V1 Admin Listings Pending Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/orders": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List Orders For Admin",
        "description": "Admin only: List all orders across the platform.",
        "operationId": "list_orders_for_admin_api_v1_admin_orders_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderRead"
                  },
                  "title": "Response List Orders For Admin Api V1 Admin Orders Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/listings/{listing_id}/approve": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Approve Listing",
        "description": "Admin only: Approve a pending listing.",
        "operationId": "approve_listing_api_v1_admin_listings__listing_id__approve_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "listing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Listing Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/listings/{listing_id}/reject": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Reject Listing Route",
        "description": "Admin only: Reject a listing (PENDING or ACTIVE).",
        "operationId": "reject_listing_route_api_v1_admin_listings__listing_id__reject_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "listing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Listing Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get My Orders",
        "description": "Get current user's orders (as buyer or seller).",
        "operationId": "get_my_orders_api_v1_orders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrderRead"
                  },
                  "type": "array",
                  "title": "Response Get My Orders Api V1 Orders Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Create Direct Order",
        "description": "Buy Now endpoint với race condition protection.",
        "operationId": "create_direct_order_api_v1_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderDirectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/orders/{order_id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "Get Order",
        "operationId": "get_order_api_v1_orders__order_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConditionGrade": {
        "type": "string",
        "enum": [
          "brand_new",
          "like_new",
          "good",
          "fair",
          "poor"
        ],
        "title": "ConditionGrade"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ListingRead": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 500,
            "minLength": 5,
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)",
            "title": "Price"
          },
          "is_negotiable": {
            "type": "boolean",
            "title": "Is Negotiable",
            "default": true
          },
          "condition_grade": {
            "$ref": "#/components/schemas/ConditionGrade"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "seller_id": {
            "type": "string",
            "format": "uuid",
            "title": "Seller Id"
          },
          "status": {
            "$ref": "#/components/schemas/ListingStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "title",
          "price",
          "condition_grade",
          "category_id",
          "id",
          "seller_id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ListingRead"
      },
      "ListingStatus": {
        "type": "string",
        "enum": [
          "pending",
          "active",
          "sold",
          "hidden",
          "rejected"
        ],
        "title": "ListingStatus"
      },
...
      "OrderDirectCreate": {
        "properties": {
          "listing_id": {
            "type": "string",
            "format": "uuid",
            "title": "Listing Id"
          },
          "use_escrow": {
            "type": "boolean",
            "title": "Use Escrow",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "listing_id"
        ],
        "title": "OrderDirectCreate"
      },
      "OrderRead": {
        "properties": {
          "final_price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d{0,2}0*$",
            "title": "Final Price"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "buyer_id": {
            "type": "string",
            "format": "uuid",
            "title": "Buyer Id"
          },
          "seller_id": {
            "type": "string",
            "format": "uuid",
            "title": "Seller Id"
          },
          "listing_id": {
            "type": "string",
            "format": "uuid",
            "title": "Listing Id"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "final_price",
          "id",
          "buyer_id",
          "seller_id",
          "listing_id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "OrderRead"
      },
      "OrderStatus": {
        "type": "string",
        "enum": [
          "pending",
          "completed",
          "cancelled",
          "disputed"
        ],
        "title": "OrderStatus"
      },
...
      "UserMe": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "province": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Province"
          },
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          },
          "ward": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ward"
          },
          "address_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Detail"
          },
          "is_phone_verified": {
            "type": "boolean",
            "title": "Is Phone Verified"
          },
          "is_email_verified": {
            "type": "boolean",
            "title": "Is Email Verified"
          },
          "trust_score": {
            "type": "number",
            "title": "Trust Score"
          },
          "rating_avg": {
            "type": "number",
            "title": "Rating Avg"
          },
          "rating_count": {
            "type": "integer",
            "title": "Rating Count"
          },
          "completed_orders": {
            "type": "integer",
            "title": "Completed Orders"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name",
          "is_phone_verified",
          "is_email_verified",
          "trust_score",
          "rating_avg",
          "rating_count",
          "completed_orders",
          "role",
          "is_active",
          "created_at",
          "updated_at"
        ],
        "title": "UserMe",
        "description": "Full user info returned to the authenticated user (GET /users/me)."
      },
...
      "UserStatusUpdate": {
        "properties": {
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "is_active"
        ],
        "title": "UserStatusUpdate",
        "description": "Schema for updating user active status by Admin."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}